Fixed issues caused by merging
The Shoot() methods we both created were both kept in code at the same time for some reason. I kept my Shoot() method in, because it's closer to what we need, and it's more intuitive. I did keep the SpriteDirection() methods, those are good, though I modified the ShootDirection() method a small amount.